refactoring (step2) : PinType 불필요한 부분 삭제
